ElfSection is no longer cheap - don't pass-by-value
ElfSections can now uncompress data on creation, so they are potentially
expensive. Also, the data length we're interestest in is independent of
sh_size, so we have to deal with that. So, stop pretending we can use
Elf_Shdr directly at all, and always use shared-pointers-to ElfSection
when we need to deal with sections.