Deal with stacks with invalid IP
In the event that we can't find the IP for the first frame on a stack,
we assume that it's the result of a bad function-call pointer - the call
instruction will have pushed the old IP on to the stack, and moved the
stack pointer down. We can create a stack frame that's a copy of the
machine register state with that instruction pointer popped off the stack.