Don't do sloppy checks against instruction pointers
Return addresses on the stack represent the instruction just past a
call being made. In a number of places, we need to extend the range we
check instruction addresses by a single byte (or subtract one from the
IP).

In order to do this consistently, we have scopeIP and rawIP for a frame.
The "scopeIP" gives the instruction pointer to use for looking up
information for that frame. For the top frame, where IP is from the CPU
register, use it directly. For other frames, we subtract one from it.

All range checks can now strictly check the scopeIP against the
instruction range they deal with.