Purge raw dies when evicting a unit from the cache.
When a unit gets chucked out of the LRU cache, purge it's RawDIE set.
Also, pass the parent DIE around the iterator code, rather than trying to
work it out when we need to set its "nextSibling" field after decoding
all the direct children. Be sure to deal with the fact that we can now
have a valid parent offset in a DIE, but the raw DIE may have been purged,
and needs to be reloaded.