Fix readsleb128 for very long values
Shifting more than the number of bits in a value has undefined behaviour,
so don't do that.  For a signed LEB value, we have 7 bits of data per
byte, so a 64-bit value will actually be encoded as 70 bits. If there
are more bits to shift than there are bits in the value, we can assume
the sign bit is already set, so do nothing.

Also, fix the name of "isSigned" - it's just the msb. And make "shift" a
reference, so it's the same as msb, and easier to read the code.