Introduce an mmap-based reader.
for ELF files, this skips the read operations and the CachedReader.
With readobj, RVO means that this generally ends up as a straight
memcpy straight from the mapped data to the copy, so the savings are
decent. (I've checked this is so for things like elfHeader/Elf32_Ehdr,
and section headers, but it's not for Phdrs). This also speeds up
"readstring" quite a bit for mmap'd files.