do full search for units iff there are no aranges.
Previously, if the search for a unit holding a given address using
aranges found nothing, we'd do an exhaustive search. This is really slow
and can be quite memory intensive.

Instead, check if the image has debug_aranges, and if so, don't second
guess a failed search by trawling the DIE tree.