Allow for a subset of units in debug_aranges
clang does not produce debug_aranges output for translation units. When
mixed with a compiler that does (eg, rustc), we get incomplete arange
info.

Instead of trying to fall back always and trawl the entire dwarf data for
an elf object, we cache DW_AT_range attributes on DW_TAG_compile_unit's,
and add them to the data we hold on to for aranges.