Let canal work on live processes, too.
This adds Process::load to create a new process from a string ID, and
uses the same logic from pstack.c to work out if the ID is a pid or a
filename (if its numeric, and you can kill it with kill(pid, 0), then
it's a process, otherwise, it's a filename). This new function is used
in pstack and in canal.

We also add an addressSpace() function, so we can get a list of the valid
address ranges in the process. This is a little ad hoc - for core files,
we just dump the LOAD sections. For live processes, we walk /proc/pid/maps

canal now uses this in preference to it's own walk of the core file
segments, and can, as a result, do its work on live processes.