Keep processes suspended while we are doing a backtrace
Although we previously got the backtrace while suspended, and then
released the process while we walked the debug data to calculate function
names, and namespaces, etc, we currently have to read process memory
to get the values of arguments, etc. You can only rely on that when the
process is still suspended from when you took the original backtrace.

We can maybe optimize for backtracing when we don't have to print
arguments.