Remove unit cache.
... and other cleanups. Now that both DIEs and abbreviations can be
purged from units, there's not much need to discard the placeholder for
the unit. This helps keep track of the offsets that units are at too,
which makes finding DIEs specified by a global offset much faster.

We still sometimes need to iterate over all units when finding a DIE
- this happens when a DIE s referenced in a different CU or a totally
different image (see DW_FORM_GNU_ref_alt) . In this case, we don't have a
PC value, just an offset in the debug_info section, and we have to find
the start of the associated CU. Parsing .gdb_index could avoid this,
if present, I think.

Get rid of "commonSections"

Just lazily evaluate bits in sections as needed. We generally only need
the name once anyway. This also moves the hash and symbol version parsing
out of Elf::Object, so you don't take the cost if you don't care about
symbol tables and/or versions.