Cleanup for stackframes
Don't use raw new/delete for stackframes. Store them as a vector of
objects, rather than pointers, and adjust some methods to match.
Refactor unwinding functions to deal with the change.
Remove the "operator->" in JSON code - it just leads to confusing
errors, and is a poor use of operator overloading.