Fix "ranges"
Previously, ranges were provided as a conversion operator from a dwarf
attribute, that was used exclusively on the DW_AT_ranges attribute.

However, ranges are also affected by the DW_AT_lowpc value, and this
distinction was lost when calculating the aranges data if it's not
provided by the compiler, leading to no functioning DIE-finding for
"go". Also, ranges can have a "base address selection" entry that was
not handled - go, of course, uses these too.

Fix both issues - ranges are now fetched for a DIE, rather than an
attribute. This means we don't have to separately manage the base
address for every range list. We still cache ranges in the compilation
unit based on the offset in the range section.