Delay creating readers a bit
Previously, all the readers for all the dwarf sections were created
up-front. This means that you must, for example, decompress things
immediately.

Recognising that creating readers is slow, we instead hold on to the
Section objects instead.

We also move the detection of the ZLIB compressed ".zdebug_..." sections
into the section's io() method, and imbue the sections with their name
as a string for ease of use,