Don't deal with "UNKNOWN" in searches.
If there is no address range information for a DIE, don't search for
address information under it. This means we can possibly not find DIEs
for functions, but we already assume that the unit in question will have
some form of address range info, so assuming the child dies do too is
not a big stretch, and seems to hold for any debuginfo I can find.