Set the IPREG if the RAR != IPREG
When everything else is unwound, set the program counter register to the
content of the RAR register. In theory, we might not have a well defined
program counter, and should just have a "return address" in the stack
frame, but I don't think that's true for any arch we'll ever care about.