Use std::map::upper_bound member function, not std::upper_bound
The former is O(log(N)), the latter is O(N).

Large shared libraries with lots of compilation units/partial
compilation units make for terrible performance when finding a DIE via
DW_FORM_ref_addr or DW_FORM_GNU_ref_alt (where the offsets are relative
to the dwarf info rather than a specific unit, and we use the upper
bound to narrow our search for the unit containing that entry.