Refactor StackFrame, and provide a "self" process type.
The intent is to provide introspection on the current process through
the libdwelf interface.

For pstack, the StackFrame type was the main interface using code
locations, but building a stack frame to find information about a
particular function is ugly.

This splits the StackFrame up, so a stack frame contains just the set of
registers, and the CFA for the frame. It also allows you to get a copy
of a new "ProcessLocation" for the instruction pointer. This contains
the ELF relocation address of the function, and a pointer to a
CodeLocation, which provides information about a specific code address
in an ELF file. That allows you to find the DIE, source information,
symbol data, etc, for that location.

With this interface, we can trivially look up symbols for code in our
own process, and find code from symbolic names, etc.

We add a simple "self" process type to allow this.

Some other changes include using the prstatus register set type intead
of converting to/from an array of registers ordered by DWARF register
assignment.