Have process subclasses override getAUXV(), not load()
We were delegating the work of load() to subclasses of process, but the
only thing each subclass needs to do differently is to find the aux
vector.

Create a new abstract method for that, and make load non-virtual, so
each subclass can just load its aux vector. This is also useful as an
API call.