Suspend LWPs, then threads, rather than vice versa
By doing this, we ensure that at the point we start enumerating threads,
the LWPs for those threads are already stopped, so the threads are not
racing us to modify the thread list as we read it. This stops a
potential hang where we end up in a loop iterating over a cycle in the
thread structures for already removed threads.