Require an explicit "undefined" register state to terminate unwinding.
We previously assumed any register that was not mentioned by the unwind
instructions would simply get its old value propagated to the next
frame, but marked their disposition as "UNDEF".

On intel, we always have to load the return address off the stack, so if
it is defined, it is referred to in the unwind instructions.

ARM uses the link register directly as the RAR, and for simple calls,
its not on the stack. It does not therefore need any instructions to
unwind it as you return, so there may be no instructions releated to the
RAR at all - the return address comes directly from the existing value
in the link register, a.ka. the RAR. Extant libc versions just don't
make any mention of the link register other than saying it's the RAR.
Our unwind process would terminate on such frames, which is obviously
incorrect.

Rejig how we treat UNDEF vs ARCH. Now, ARCH is the default state of a
register that is not mentioned in the unwind instructions. We require
the RAR to be explicitly UNDEF to terminate unwinding. This seems fine
for intel archs, too.