Add "freeres" option for debugging
Make the call to exit before the ImageCache leaves scope conditional on
a new flag, so we can check that everything is properly deleted if
we're running under some tool like valgrind.

Use this as an excuse to make flags.{cc,h} work for options that don't
have a short-from version. (There's so many flags now that ths single
new short options are mostly just confusing)