Allow fetching of FP registers
Generalize the getRegs() call so that you can ask for a specific regset.
For live processes, use ptrace, and for dead processes, find the
accompanying note.

For the core processes, we need to clean up the way we handle notes in
general - we just store a vector of note descriptors, which each has a
reader. We also store the indexes in the array of the NT_PRSTATUS notes,
so we can find the starting point for a particular thread when we need
it.

Remove the LogProcess as a failed experiment