Implement support for eh_frame_hdr section.
Previously, we scanned all the FDEs and CIEs when creating the call
frame information for an ELF object.

First, replace the list<FDE> with a vector<unique_ptr<FDE>>, and now
sort the FDEs by the address of the first instruction they cover (the
iloc field) once we're done, and sort when we're done. This allows
findFDE to search the table using a binary search based on the requested
function pointer value.

The previously ignored .eh_frame_hdr includes a table giving, for each
FDE in .eh_frame, the instruction pointer of the start of the
instruction range covered, and the offset in the eh_frame section of the
FDEs header. This is sorted by the iloc value also, so should be in the
same order as our table above.

Therefore, if we do have an eh_frame_hdr section, we don't need to load
all the FDEs - we just create an empty vector, and, when binary
searching, if the unique_ptr<FDE> is null for an entry, we can load it
from the eh_frame_hdr table at the same index as our own fdes table, and
populate our FDEs table when we're done.