arm: if we are looking at the signal trampoline, it's scopeIP == rawIP
On Aarch64, there's no FDE for the signal trampoline, so we detect it by
finding the name of the symbol. In this case, we have a specific flag in
the CodeLocation to indicate this is a signal trampoline, as we have no
CIE to indicate it. If we have such a flag in the stack frame, don't
bother looking for the FDE/CIE, and just return the raw IP address for
scopeIP (there's no need to subtract 1 to get into the call instruction,
as there was no call instruction, and subtracting one will bring us out
of the function itself in this case)

This is mostly cosmetic - we unwind properly through this frame, but we
don't find the symbol for it correctly.