Fix FDE printing
Iterating over all the FDEs in the dump was broken by the lazy eval of
FDEs. This is a quick-and-dirty fix.

We need to make the cie and fde sets mutable, which is in-keeping with
other similar lazy caches, and, for now, call "ensureFDEs" to make sure
all FDEs are available. It would be better to have a proper
iterable/iterator type that would do this behind the scenes.