Add verbatim CPython 3.9's dict-common
This patch brings a header file from the CPython source that defines the
implementation details of a dictionary, which is correct for Python
versions up to 3.11. Previously, the header was referenced directly with
the `#include <Objects/dict-common.h>` directive, with the assumption
that CPython's internals would be present somewhere in the include
directories.

With Python 3.11, the location of the definitions previously provided by
`dict-common.h` has been moved to `Internal/*`, which has broken the
builds on recent distro releases.

Since we don't differentiate between minor versions of Python when
reading the stack (the CPython source present on the build host
determines how Python dictionaries will be read), we may as well include
the header containing dict layout information here, directly in pstack.

Pstack uses a BSD-2-clause license, which is compatible with CPython's
PSFL.

The CPython version from which the header was copied is 3.9.21.