Fix !ZLIB compilation
Starting with af59ef83 the builds without WITH_ZLIB would fail due to
`debug` not being defined.
The `Section` object has a reference to the
elf it belongs to, and that elf has the context with the debug stream.
This patch makes use of that relationship to push the debug messages.