Fix execInsns
This was previously a recursive function - it was called once for an
FDE, and the first call would recurse to execute the CIE's instructions
to get the default call frame state, then continue with the FDEs
instructions up to the desired instruction pointer.

We now have a common function template, that takes a function that can
yield the execution of instructions before the end of the sequence, and
a function on the FDE, that calls this, first with the full stream from
the CIE, then with the local FDE stream, this time with a yielding
function that will stop when the desired instruction pointer is reached.

In theory, we can use this to do better caching (eg, when calculating
the call frame for a specific instruction pointer we can start at the
previously evaluated instruction pointer, rather than from the start of
the FDE again), and I also want to play with looking at the instructions
without having to set up the plumbing to unwind the stack.

Clean up some naming, too.