More complete debuginfod support
This makes the "Context" layer responsible for all the build-id vs path
stuff. You can get an image by ether, and they are cached appropriately.

Both executables and debug images are fetched preferentially by
build-id. We use the local file system as the primary source, then
debuginfod, finally falling back to name-based lookup when all else
fails.

We also look at the PHDRs in the auxv to find the build-id of the
executable