Fix some issues with "AddressSpace"s in core processes.
Populate the "offset" from the start of a file, based on the File note,
and set the inode to be something other than 0 in the case that the file
exists.

Also, because the FileEntry type is only used by the "dead" process,
move it into the .cc file to stop exposing the public headers to it.

Also, previously, vmflags and permissions were aggregated into one
field. After splitting, the local variable used to build up the
permissions remained called "flags". Fix that, and rename it.