implement thread names for live processes, and improve JSON utilities
The first part is pretty easy - just use the name of the task from
/proc/<pid>/task/stat.

Actually making the test work was the straw that broke the camel's back
WRT how the JSON formatting works, so fix that up:

*  make all our iterators obey the full concept that is required to make
   them proper input-iterators.

*  Make the containers proper std::range::ranges, and use that
   consistently as the idiom for things that print containers

*  clean up some code by using concepts, and make the difference between
   maps and objects be that objects have value types that are pairs
   where `first` is a string-like thing

*  remove the horrible handrolled "Mapper" mess, and use
   std::views::transform to create transformed views when necessary.

Then change the test infra so we can use pstack on processes as well as
cores, and the "child" mode of operation that exists only for testing.
Only the "live" mode will support thread names - the required
information just isn't present in a core.

This should fix issue #39