Don't assume we support all the registers on a platform
The debuginfo can contain references to registers that we may not yet
support - for example, for aarch64, we don't parse out the additional
registers in the __reserved field of the mcontext_t.

So, don't use map::at to access such registers, just use find, and
ignore the case where we don't know anything about the register.