Don't assume we can restore every register.
We only deal with a limited subset of DWARF registers - namely the
"general purpose" registers, or those in the user_regs_struct. for
i386/x86_64, that's pretty much the only thing that appears in unwind
info - the floating point and SIMD registers are all caller-saved, so
the unwind info never includes info about how to restore them on unwind.
That is not true for ARM, so we need to be careful when we get

instructed to restore a register from a saved blob - we may not have
existing state for that register.

I have a proper fix for this in the main branch - it revamps the way we
handle registers entirely, so we can deal properly with floating point
and SIMD registers, but it's not finished, so doing this quick hack for
now.