Hail mary stack unwwind for ARM
On ARM, if we have no CFI, and no other option, we can just look at the
link register, and replace the instruction pointer with it. Unlike on
32- and 64-bit intel platforms, it appears that the linux aarch64
toolchain will not generate FDE data for PLT entries - this means we
need this for the case that a process happens to be in a PLT entry when
we take a backtrace.