Reduce support for thread_db to the barest minimum
The only practical use at this point is to give pthread thread ids for
the threads, so just do that, and make everything else rely on LWPs
everywhere. We can also now get our full CoreRegisters when unwinding
the stack.

Change the register get interfaces to throw exceptions when the fail -
it's too easy to silently ignore at the moment.